vim syntax: Allow "x86_64_v3" for $arch
Arch Linux RFC 2[1] added "x86_64_v3" as a (future) official Arch Linux
build architecture, but vim is currently not happy when `PKGBUILD`s have
"x86_64_v3" in their `$arch` array.

This adds "x86_64_v3" to the list of allowed `pbArch` keywords as
well as the various `{source,depends,…}_$arch` to the list of
`srcInfoField`s for `.SRCINFO`.

[1] https://rfc.archlinux.page/0002-march/