Revert 'printf' lint for the pacscripts script
The lint change made in commit [17f27b53](!33 (17f27b53)) for the `pacscripts`
script regarding the `printf` statement broke the expected formatting.

This parallels a previous [issue](#118) with the `updpkgsums` script,
which was fixed in commit [88c4b490](!38 (88c4b490)).

As far as I can tell, we have an exception to shellcheck rule SC2059,
because the first argument to `error()` is meant to be a format string
that is filled with remaining arguments. Interestingly enough, this
script does not even use the `error()` function this way, but calls it
with single strings built by variable expansion.